# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "The Reward Loyalty API provides programmatic access to the Reward Loyalty platform, enabling integration with external systems, mobile apps, and custom workflows. This RESTful API supports authentication, loyalty card management, member operations, partner operations, and transaction processing.\n\n## Authentication\n\nAll authenticated endpoints use Bearer token authentication via Laravel Sanctum. Include your token in the Authorization header.\n\nTokens are obtained by calling the login endpoints for each user type (Admin, Partner, or Member).\n\n## Rate Limiting\n\nAPI requests are rate-limited to ensure fair usage. Standard limits apply per authenticated user.\n\n## Response Format\n\nAll responses are returned in JSON format. Successful responses include the requested data, while errors include a message field describing the issue.",
        "contact": {
            "name": "Reward Loyalty Website",
            "url": "https://distech.co.za"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Current Server"
        }
    ],
    "paths": {
...
        "/{locale}/v1/member/stamp-cards": {
            "get": {
                "tags": [
                    "Member"
                ],
                "summary": "Retrieve member's stamp cards",
                "description": "Retrieve all stamp cards the authenticated member is enrolled in, including current progress, stamps collected, and rewards earned. Also returns aggregate statistics.",
                "operationId": "getMemberStampCards",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stamp cards retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StampCardsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "member_auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "description": "Response when authentication is required but not provided or invalid",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Authentication error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "StampCard": {
                "description": "A stamp-based loyalty card",
                "properties": {
                    "id": {
                        "description": "Unique identifier",
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "description": "Card title",
                        "type": "string"
                    },
                    "description": {
                        "description": "Card description",
                        "type": "string",
                        "nullable": true
                    },
                    "reward_title": {
                        "description": "Reward title",
                        "type": "string"
                    },
                    "reward_description": {
                        "description": "Reward description",
                        "type": "string",
                        "nullable": true
                    },
                    "current_stamps": {
                        "description": "Member's current stamp count",
                        "type": "integer"
                    },
                    "stamps_required": {
                        "description": "Stamps needed to complete",
                        "type": "integer"
                    },
                    "progress_percentage": {
                        "description": "Completion progress (0-100)",
                        "type": "number",
                        "format": "float"
                    },
                    "pending_rewards": {
                        "description": "Unclaimed completed rewards",
                        "type": "integer"
                    },
                    "completed_count": {
                        "description": "Total times completed",
                        "type": "integer"
                    },
                    "redeemed_count": {
                        "description": "Total rewards redeemed",
                        "type": "integer"
                    },
                    "enrolled_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "last_stamp_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "last_completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "next_stamp_expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "stamp_icon": {
                        "description": "Icon identifier",
                        "type": "string",
                        "nullable": true
                    },
                    "colors": {
                        "description": "Card color scheme",
                        "type": "object"
                    },
                    "media": {
                        "description": "Card media assets",
                        "type": "object"
                    },
                    "settings": {
                        "description": "Card behavior settings",
                        "type": "object"
                    },
                    "reward_value": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "reward_points": {
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "StampCardsResponse": {
                "description": "Response containing member's stamp cards and statistics",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "properties": {
                            "stamp_cards": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/StampCard"
                                }
                            },
                            "stats": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "admin_auth_token": {
                "type": "http",
                "description": "Administrator authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "partner_auth_token": {
                "type": "http",
                "description": "Partner authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Member authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}